Added dialogue sfx

+++ Dateformer/Assets/Scripts/Dialogue/SceneScript.cs
@@ -28,7 +28,7 @@ public class SceneScript : MonoBehaviour
     {
         yield return new WaitForSeconds(1);
         character.dialogue.Say(characterLines[dialogueIndex].dialogue, character.characterName);
-        character.SetSprite((int)characterLines[dialogueIndex].currentMood);
+        character.SetSprite(characterLines[dialogueIndex].currentMood);
         startedDialogue = true;
         if (characterLines[dialogueIndex].promptForResponse)
         {
@@ -64,7 +64,7 @@ public class SceneScript : MonoBehaviour
                     return;
                 }
                 character.Say(characterLines[dialogueIndex].dialogue);
-                character.SetSprite((int)characterLines[dialogueIndex].currentMood);
+                character.SetSprite(characterLines[dialogueIndex].currentMood);
                 if (characterLines[dialogueIndex].promptForResponse)
                 {
                     choiceIndex = characterLines[dialogueIndex].choiceIndex;
@@ -85,7 +85,7 @@ public class SceneScript : MonoBehaviour
         character.dialogue.isWaitingForResponse = false;
         character.dialogue.CloseChoicePanel();
         character.Say(characterLines[dialogueIndex].dialogue);
-        character.SetSprite((int)characterLines[dialogueIndex].currentMood);
+        character.SetSprite(characterLines[dialogueIndex].currentMood);
 
         if (characterLines[dialogueIndex].promptForResponse)
         {
@@ -104,7 +104,7 @@ public class SceneScript : MonoBehaviour
         character.dialogue.isWaitingForResponse = false;
         character.dialogue.CloseChoicePanel();
         character.Say(characterLines[dialogueIndex].dialogue);
-        character.SetSprite((int)characterLines[dialogueIndex].currentMood);
+        character.SetSprite(characterLines[dialogueIndex].currentMood);
 
         if (characterLines[dialogueIndex].promptForResponse)
         {
